requirements.txt
streamlit
Pillow
requests
streamlit-autorefresh
msal